Extract fragment related code into separate file
