+++ server.go
@@ -86,31 +86,6 @@ func NewServer(target string) *Server {
 	}
 }
 
-type ContentFragments = []*fragments.Definition
-type DefinitionOption = fragments.DefinitionOption
-
-func WithFragmentMetadata(metadata map[string]string) DefinitionOption {
-	return fragments.WithMetadata(metadata)
-}
-
-// DefineFragment creates a new fragments.Definition targetting the given path.
-func DefineFragment(path string, opts ...DefinitionOption) *fragments.Definition {
-	return fragments.New(path, opts...)
-}
-
-// DefineFragments creates a new []*fragments.Definition for each string in
-// paths. This can be used in combination with the Server.Get function to define
-// routes.
-func DefineFragments(paths ...string) ContentFragments {
-	definitions := make(ContentFragments, len(paths))
-
-	for i, path := range paths {
-		definitions[i] = DefineFragment(path)
-	}
-
-	return definitions
-}
-
 type GetOption = func(*Route)
 
 func WithRouteMetadata(metadata map[string]string) GetOption {
